% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paquete_R.R
\name{clase_pendiente}
\alias{clase_pendiente}
\title{Clasificar la pendiente a partir de un DEM}
\usage{
clase_pendiente(dem, unit = "degrees")
}
\arguments{
\item{dem}{Un objeto \code{terra::SpatRaster} que representa un DEM.}

\item{unit}{Unidad de la pendiente. Por defecto \code{"degrees"}.}
}
\value{
Un \code{terra::SpatRaster} con clases de pendiente.
}
\description{
Esta función calcula la pendiente desde un modelo digital de elevaciones (DEM)
y la clasifica en cuatro categorías:
1 = llana, 2 = suave, 3 = moderada, 4 = fuerte.
}
