% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipos.R
\docType{data}
\name{equipos}
\alias{equipos}
\title{Estadísticas de los equipos de la Liga Nacional de Básquet Argentina}
\format{
Un data frame con 2280 filas y 30 variables:
\describe{
\item{\code{id}}{Identificador del partido (variable de soporte).}
\item{\code{source}}{Código de referencia utilizado durante el web scraping (variable de soporte).}
\item{\code{Equipo}}{Equipo que disputó el partido.}
\item{\code{Min}}{Minutos jugados por el equipo en el partido.}
\item{\code{2Con}}{Dobles convertidos por el equipo en el partido.}
\item{\code{2Int}}{Dobles intentados por el equipo en el partido.}
\item{\code{3Con}}{Triples convertidos por el equipo en el partido.}
\item{\code{3Int}}{Triples intentados por el equipo en el partido.}
\item{\code{FG\%}}{Porcentaje de tiros de campo convertidos por el equipo en el partido.}
\item{\code{1Con}}{Tiros libres convertidos por el equipo en el partido.}
\item{\code{1Int}}{Tiros libres intentados por el equipo en el partido.}
\item{\code{1\%}}{Porcentaje de tiros libres convertidos por el equipo en el partido.}
\item{\code{Ro}}{Rebotes ofensivos ganados por el equipo en el partido.}
\item{\code{Rd}}{Rebotes defensivos ganados por el equipo en el partido.}
\item{\code{Reb}}{Rebotes totales ganados por el equipo en el partido.}
\item{\code{Ast}}{Asistencias realizadas por el equipo en el partido.}
\item{\code{Bp}}{Pérdidas de balón cometidas por el equipo en el partido.}
\item{\code{Br}}{Recuperaciones de balón realizadas por el equipo en el partido.}
\item{\code{Tap}}{Tapas realizadas por el equipo en el partido.}
\item{\code{Fa}}{Faltas personales cometidas por el equipo en el partido.}
\item{\code{Pts}}{Puntos convertidos por el equipo en el partido.}
\item{\code{Val}}{Valoración asignada al equipo en el partido.}
\item{\code{Temp}}{Temporada a la que corresponde el partido.}
\item{\code{df}}{Código de referencia utilizado durante el web scraping (variable de soporte).}
\item{\code{Condicion}}{Condición en la que el equipo disputó el partido.}
\item{\code{Rival}}{Rival del equipo que disputó el partido.}
\item{\code{PtsRec}}{Puntos recibidos por el equipo en el partido.}
\item{\code{Resultado}}{Resultado que obtuvo el equipo en el partido.}
\item{\code{Dif}}{Diferencia de puntos en el partido entre el equipo y su rival.}
\item{\code{Estado}}{Estado y disponibilidad de los datos del partido.}
}
}
\source{
\href{https://www.proballers.com/es}{Proballers}
}
\usage{
equipos
}
\description{
Dataset con información de rendimiento por equipo en cada partido.
}
\keyword{datasets}
