% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jugadores.R
\docType{data}
\name{jugadores}
\alias{jugadores}
\title{Estadísticas de los jugadores de la Liga Nacional de Básquet Argentina}
\format{
Un data frame con 22192 filas y 32 variables:
\describe{
\item{id}{Identificador del partido (variable de soporte).}
\item{source}{Código de referencia utilizado durante el web scraping (variable de soporte).}
\item{Jugador}{Jugador que disputó el partido.}
\item{Min}{Minutos jugados por el jugador en el partido.}
\item{2Con}{Dobles convertidos por el jugador en el partido.}
\item{2Int}{Dobles intentados por el jugador en el partido.}
\item{3Con}{Triples convertidos por el jugador en el partido.}
\item{3Int}{Triples intentados por el jugador en el partido.}
\item{FG\\%}{Porcentaje de tiros de campo convertidos por el jugador en el partido.}
\item{1Con}{Tiros libres convertidos por el jugador en el partido.}
\item{1Int}{Tiros libres intentados por el jugador en el partido.}
\item{1\\%}{Porcentaje de tiros libres convertidos por el jugador en el partido.}
\item{Ro}{Rebotes ofensivos ganados por el jugador en el partido.}
\item{Rd}{Rebotes defensivos ganados por el jugador en el partido.}
\item{Reb}{Rebotes totales ganados por el jugador en el partido.}
\item{Ast}{Asistencias realizadas por el jugador en el partido.}
\item{Bp}{Pérdidas de balón cometidas por el jugador en el partido.}
\item{Br}{Recuperaciones de balón realizadas por el jugador en el partido.}
\item{Tap}{Tapas realizadas por el jugador en el partido.}
\item{Fa}{Faltas personales cometidas por el jugador en el partido.}
\item{Pts}{Puntos convertidos por el jugador en el partido.}
\item{Val}{Valoración asignada al jugador en el partido.}
\item{Equipo}{Equipo en el que juega el jugador.}
\item{PtsEqui}{Puntos convertidos por el equipo del jugador.}
\item{Temp}{Temporada a la que corresponde el partido.}
\item{df}{Código de referencia utilizado durante el web scraping (variable de soporte).}
\item{Condicion}{Condición en la que el jugador disputó el partido.}
\item{Rival}{Rival del jugador que disputó el partido.}
\item{PtsRec}{Puntos recibidos por el equipo del jugador en el partido.}
\item{Resultado}{Resultado que obtuvo el jugador en el partido.}
\item{Dif}{Diferencia de puntos en el partido entre el equipo del jugador y su rival.}
\item{Estado}{Estado y disponibilidad de los datos del partido.}
}
}
\source{
\href{https://www.proballers.com/es}{Proballers}
}
\usage{
jugadores
}
\description{
Dataset con información de desempeño individual por partido.
}
\keyword{datasets}
