% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{ndiff}
\alias{ndiff}
\title{Normalized difference tool}
\usage{
ndiff(x, y)
}
\arguments{
\item{x}{Un objeto con valores numéricos}

\item{y}{Un objeto con valores numéricos de la misma clase que x}
}
\value{
valores numéricos del índice
}
\description{
\code{ndiff} calcula el índice de diferencia normalizada
}
\examples{
ndiff(c(0.8,0.6),c(0.05,0.06))
ndiff(0.1,0.09)
}
