% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial.R
\name{sig_buffer}
\alias{sig_buffer}
\title{Buffer GIS tool}
\usage{
sig_buffer(capa_entrada, distancia, dissolve = TRUE, campo)
}
\arguments{
\item{capa_entrada}{Un objeto vectorial de la clase sf}

\item{distancia}{Distancia de vecindad para determinar la zona de influencia (m)}

\item{dissolve}{Parametro para controlar la fusion de geometrias}

\item{campo}{Parametro para fusionar por un atributo. Opcional}
}
\value{
Un objeto vectorial de la clase sfc_POLYGON
}
\description{
\code{sig_buffer} devuelve una zona de influencia entorno a las geometrias de entrada
}
\examples{
\dontrun
sig_buffer(parcelas, 1000)
sig_buffer(parcelas, 1000, dissolve = T)
sig_buffer(parcelas, 1000, dissolve = T, 'Especie')

}
